cortex-ai-develop/requirements.txt
numpy 
pandas
scipy
sklearn
nltk
gensim
seaborn
matplotlib
textblob
requests
bs4
instafetch
jupyter
firefly-python
